% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asterisks.R
\name{asterisks}
\alias{asterisks}
\title{Asterisks function}
\usage{
asterisks(p_value)
}
\arguments{
\item{p_value}{p value}
}
\value{
Returns the p_value in asterisks
}
\description{
Help function to convert p_value to asterisks
}
