% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassExercise-Birthday.r
\name{birthday}
\alias{birthday}
\title{Birthday Problem Probability}
\usage{
birthday(n)
}
\arguments{
\item{n}{Integer or vector of integers. The number of people in the group(s).}
}
\value{
A numeric vector of probabilities for each value of \code{n}.
}
\description{
Calculates the probability that at least two people in a group of \code{n}
share the same birthday, assuming 365 days in a year.
}
\examples{
birthday(n = 20:25)
}
