% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Linear Regression Coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{object of class "linreg"}

\item{...}{Parameters}
}
\value{
Regression coefficient
}
\description{
From S3 function return coefficient object
}
