% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_linreg.R
\name{coef}
\alias{coef}
\title{S3 Generic function for coefficient \code{linreg}}
\usage{
coef(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
Coefficients method of S3 class
}
