% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_code.R
\name{graph_code}
\alias{graph_code}
\title{Code and Info for Graphs}
\usage{
graph_code(df, type, x, y = NULL, color = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{type}{The name (in quotes) of the graph of interest.}

\item{x}{The first variable of interest (will go on the x axis of the graph)}

\item{y}{An additional variable of interest (will go on the y axis of the graph)}

\item{color}{An additional variable of interest (will be used as the color argument of the graph)}
}
\value{
graph_code returns a printed output which contains the code to create the graph of interest with variable names in their correct places, along with information about when the usage of the graph of interest.
}
\description{
graph_code is an educational function used to learn about how to write code using the ggplot2 package for different types of graphs generated by the what_graph function {GGHelper}
}
\details{
It is important that before this function is used, the columns of interest have the correct data type.
Specifically, a categorical or discrete variable needs to have a data type of 'factor' and a continuous variable needs to have a data type of 'numeric'

Additionally, in order for the function to work, the variables that are input in the function need to be contained inside the data set that is input.
}
\examples{
graph_code(iris, "Scatter Plot", x = Sepal.Length, y = Sepal.Width)

}
