\name{lab4package-package}
\alias{lab4package-package}
\alias{lab4package}
\docType{package}
\title{
\packageTitle{lab4package}
}
\description{
\packageDescription{lab4package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab4package}
\packageIndices{lab4package}

}
\author{
\packageAuthor{lab4package}

Maintainer: \packageMaintainer{lab4package}
}

\keyword{ package }
