% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4package-package.R
\docType{package}
\name{lab4package}
\alias{lab4package}
\alias{lab4package-package}
\title{Linear Regression}
\description{
This linear regression package contains a reference
class. See \code{\link{linreg}} for more info.
}
\author{
\strong{Maintainer}: Amer Memic \email{ameme698@student.liu.se}

Authors:
\itemize{
  \item Olof Josefsson \email{olojo524@student.liu.se}
  \item Shahreen Saleem \email{shahreen.saleem111@gmail.com}
}

}
