% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class for linear regression}
\description{
This linear regression reference class will compute
the coefficients, residuals, predicted values, standard errors
for both the coefficients and the residuals, t-values and
p-values, and the degrees of freedom of a model matrix based
on the provided formula and data set.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula object}

\item{\code{data}}{A data set}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{Returns the coefficients.}

\item{\code{initialize(formula, data)}}{The constructor will return an object of class linreg.}

\item{\code{plot()}}{Plots two graphs; residuals vs fitted values, and scale-location.}

\item{\code{pred()}}{Returns the predicted values.}

\item{\code{print()}}{Prints how the function was called and the coefficients.}

\item{\code{resid()}}{Returns the residuals.}

\item{\code{summary()}}{Prints a summary of the model.}
}}

