% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression}
\usage{
linreg(formula, data, ...)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{a data frame}

\item{...}{additional argument \code{model.matrix} is a function}
}
\value{
List of outputs
}
\description{
The 'linreg' function calculate the linear regression model
}
\note{
'linreg' is a function like "lm" from "stats" package
}
\examples{
data("iris")
linreg(Petal.Length ~ Species, iris)
}
