% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printascii.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap estimator with histogram and confidence interval}
\usage{
myboot2(iter = 10000, x, fun = mean, alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations.}

\item{x}{Numeric vector of data.}

\item{fun}{Function (or function name as a string) to compute the statistic.}

\item{alpha}{Significance level for the confidence interval.}

\item{cx}{Character expansion factor for printed text.}

\item{...}{Additional graphical arguments passed to hist().}
}
\description{
Bootstrap estimator with histogram and confidence interval
}
\details{
Used in Lab 9 for MATH-4753.
}
\examples{
\dontrun{
  myboot2(10000, x = rnorm(20), fun = "mean", alpha = 0.05, cx = 1.5)
}

}
