% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printascii.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem simulation for the Poisson Distribution
Used to demonstrate that it'll work with continuous or discrete distributions}
\usage{
mycltp(n = 10, iter = 10000, lambda = 10, ...)
}
\arguments{
\item{n}{Integer. Sample size for each iteration. Default is 10.}

\item{iter}{Integer. Number of iterations (simulated samples). Default is 10000.}

\item{lambda}{Numeric. The rate parameter of the Poisson distribution. Default is 10.}

\item{For}{options to modify the graphical output of \code{hist()}.}
}
\value{
This function produces three plots:
\itemize{
\item A histogram of the sample means with a theoretical normal curve.
\item A barplot of the simulated Poisson samples.
\item A plot of the theoretical Poisson probability mass function.
}
}
\description{
Central Limit Theorem simulation for the Poisson Distribution
Used to demonstrate that it'll work with continuous or discrete distributions
}
\details{
Used in Lab8 MATH-4753
}
\examples{
\dontrun{
  mycltp(n = 10, iter = 10000, lambda = 10)
}

}
