% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printascii.R
\name{myhyper}
\alias{myhyper}
\title{MyHyper from Lab5
Simulates hypergeometric distribution via repeated sampling}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{r}{Integer, number of "Successes"}

\item{n}{Integer, sample size drawn without replacement}

\item{x}{Integer, determines the number of iterations}
}
\description{
MyHyper from Lab5
Simulates hypergeometric distribution via repeated sampling
}
\examples{
myhyper(iter=100,N=20,r=12,n=5)
}
