% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printascii.R
\name{myncurve}
\alias{myncurve}
\title{Modified function for lab 7
Displays the curve, shaded area between it and x axis from
neg inf to x=a, then calculates the area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu, }{Mean of the given normal distribution}

\item{sigma, }{Standard Deviation of given normal distribution}

\item{a, }{upper bound for the shaded area}
}
\description{
Modified function for lab 7
Displays the curve, shaded area between it and x axis from
neg inf to x=a, then calculates the area
}
