% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{new_linreg}
\alias{new_linreg}
\title{Linreg Class Constructor}
\usage{
new_linreg(object = list())
}
\arguments{
\item{object}{A list}
}
\value{
The list input from class 'linreg'
}
\description{
Calculate Linear regression
}
