% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot graph of Residuals vs Fitted values and Standardized residuals vs Fitted values}
\usage{
\method{plot}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\value{
Plots
}
\description{
An S3 function to plot Residuals vs Fitted graph
}
