% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linreg.R
\name{plot}
\alias{plot}
\title{Plot Residuals vs Fitted values}
\usage{
plot(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
Plot S3 class method
}
