% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_linreg.R
\name{pred}
\alias{pred}
\title{S3 Generic function for Predicted values (y) \code{linreg}}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
Predicted values (y) method of S3 class
}
