% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print coefficients}
\usage{
\method{print}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
An S3 function to print coefficients
}
