% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_linreg.R
\name{print}
\alias{print}
\title{S3 Generic function to print linreg}
\usage{
print(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
Print S3 class method
}
