% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printascii.R
\name{printascii}
\alias{printascii}
\title{Print ASCII}
\arguments{
\item{x}{String containing the filename
If valid file, is printed
File to be printed must be in working directory
Examples
print_ascii("light_ascii.txt") (Provided)}
}
\description{
More Detailed Description
}
