% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Vector of residuals}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{object of class "linreg"}

\item{...}{Parameters}
}
\value{
The residuals
}
\description{
From S3 function return residuals object
}
