% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_linreg.R
\name{resid}
\alias{resid}
\title{S3 Generic function for Vector of residuals (e) \code{linreg}}
\usage{
resid(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
Vector of residuals (e) method of S3 class
}
