% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Predicted values}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{object of class "linreg"}

\item{...}{Parameters}
}
\value{
The predicted value
}
\description{
From S3 function return predicted object
}
