% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_linreg.R
\name{summary}
\alias{summary}
\title{S3 Generic function for summary \code{linreg}}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Parameters}
}
\description{
Summary method of S3 class
}
