% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg RC Class}
\value{
Nothing.
}
\description{
This class offers linear regression calculation using the QR Decomposition method and comes with some handy tools.
}
\section{Fields}{

\describe{
\item{\code{X}}{matrix. Independent Values.}

\item{\code{y}}{matrix. Depenent Values.}

\item{\code{b_hat}}{matrix. Regression Coeffcients.}

\item{\code{y_hat}}{matrix. Fitted Values.}

\item{\code{l_e}}{matrix. Residuals.}

\item{\code{l_n}}{numeric. Number of Independant Values (Number of Observations).}

\item{\code{l_p}}{numeric. Number of Dependant Values (Number of Parameters in the Model).}

\item{\code{l_df}}{numeric. Degrees of freedom.}

\item{\code{l_sigma_s}}{matrix. Residual Variance.}

\item{\code{l_var_beta}}{matrix. Variance of the Regression Coefficients.}

\item{\code{l_formula}}{formula. The Formula for the Linear Regression.}

\item{\code{l_p_values}}{numeric. P-Values.}

\item{\code{l_t_beta}}{matrix. T-Values for each Coefficient.}

\item{\code{l_data_set_name}}{character. The Given Data.}
}}


