% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg RC Class}
\value{
Nothing.
}
\description{
This class offers linear regression calculation using the QR Decomposition method and comes with some handy tools.
}
\section{Fields}{

\describe{
\item{\code{X}}{matrix. Independent values.}

\item{\code{y}}{matrix. Dependent values.}

\item{\code{b_hat}}{matrix. Regression coefficients.}

\item{\code{y_hat}}{matrix. Fitted values.}

\item{\code{e_hat}}{matrix. Residuals values.}

\item{\code{df}}{numeric. Degrees of freedom.}

\item{\code{s2_hat}}{matrix. Variance of residuals.}

\item{\code{var_b_hat}}{matrix. Variance of the regression coefficients.}

\item{\code{f_text}}{formula. The formula for the linear regression.}

\item{\code{p_value}}{numeric. P-Values.}

\item{\code{t_b}}{matrix. T-Values for each coefficient.}

\item{\code{name_ds}}{character. The given data.}

\item{\code{Q}}{matrix. Q matrix from the QR decomposition.}

\item{\code{R}}{matrix. Upper triangular matrix.}

\item{\code{Qy}}{matrix.}

\item{\code{sterr}}{matrix. Standard error.}

\item{\code{print()}}{function. Returns coefficients.}

\item{\code{plot}}{function. Returns sets of plots..}

\item{\code{resid}}{function. Returns residuals.}

\item{\code{pred}}{function. Returns predictions.}

\item{\code{coef}}{function. Returns coefficients.}

\item{\code{summary}}{function. Returns summary of the above.}
}}


