% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_graph.R
\name{what_graph}
\alias{what_graph}
\title{Gives a list of potential graphs and charts}
\usage{
what_graph(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{n}{Number of variables to graph}

\item{var1}{A variable from dataframe}

\item{var2}{A variable from dataframe (optional)}

\item{fill}{A variable from dataframe used as the fill/color aesthetic (optional)}
}
\value{
A vector of possible graphs
}
\description{
Gives a list of potential graphs and charts
}
\examples{
what_graph(iris, 1, Sepal.Length)

}
